requirements.txt
pandas~=1.0.3
torch~=1.2.0
sklearn~=0.0
scikit-learn~=0.22
transformers~=3.0.2
numpy~=1.18.1
pathlib~=1.0.1
PyYAML~=5.1.2
attrs~=19.3.0